requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Graph ML
torch>=2.5.0
torch-geometric>=2.6.0

# Graph Analysis
networkx>=3.4

# ML / Classical
xgboost>=2.1.0
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.2.0

# LLM & RAG
google-generativeai>=0.8.0
faiss-cpu>=1.9.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# Utilities
pyyaml>=6.0
httpx>=0.28.0
loguru>=0.7.0
tqdm>=4.66.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0

# Dev
black>=24.0.0
ruff>=0.8.0
